LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/view-timeline-lookup-expected.txt


FAIL view-timeline on self assert_equals: expected "100" but got "0"
PASS timeline-scope on preceding sibling
PASS view-timeline on ancestor
PASS timeline-scope on ancestor sibling
FAIL timeline-scope on ancestor sibling, conflict remains unresolved assert_equals: expected "auto" but got "50"
FAIL timeline-scope on ancestor sibling, closer timeline wins assert_equals: expected "auto" but got "75"
PASS view-timeline on ancestor sibling, scroll-timeline wins on same element
